// frontend/src/lib/api/roborev/generated/models/review.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { ReviewJob } from "./reviewJob.ts";

export interface Review {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  agent: string;
  closed: boolean;
  created_at: string;
  id: number;
  job?: ReviewJob;
  job_id: number;
  output: string;
  prompt: string;
  synced_at?: string;
  updated_at?: string;
  updated_by_machine_id?: string;
  uuid?: string;
  verdict_bool?: number;
}