// frontend/src/lib/api/roborev/generated/models/response.ts
/**
 * Generated by Orval. Do not edit manually.
 */

export interface Response {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  commit_id?: number;
  created_at: string;
  id: number;
  job_id?: number;
  responder: string;
  response: string;
  source_machine_id?: string;
  synced_at?: string;
  uuid?: string;
}